api/requirements.txt
Django==3.2
django-grappelli
django-nested-admin==3.3.3